requirements.txt
azure-devops
pandas
python-dotenv
tqdm